Build PythonCore as a static lib.
Unfortunately Python is not robust enough to be used as a DLL. It doesn't
properly clean up after itself. It carries too much global state.

To ensure the global state is cleaned up after each use, we need to
statically link it against an auxiliary DLL which embeds Python and is
unloaded after use.

This change makes the necessary changes to allow for static linking.